XRT-1320 Support added for AzureSphere SDK 22.02 and API 12

+++ src/c/threadpool.c
@@ -160,7 +160,7 @@ iot_threadpool_t * iot_threadpool_alloc (uint16_t threads, uint32_t max_jobs, in
   iot_threadpool_t * pool = (iot_threadpool_t*) calloc (1, sizeof (*pool));
   pool->affinity = affinity;
   pool->logger = logger;
-  *((uint16_t*) &pool->id) = atomic_fetch_add (&pool_id, 1u);
+  *((uint16_t*) &pool->id) = (uint16_t) atomic_fetch_add (&pool_id, 1u);
   iot_logger_add_ref (logger);
   iot_log_info (logger, "iot_threadpool_alloc (threads: %" PRIu16 " max_jobs: %u default_priority: %d affinity: %d)", threads, max_jobs, default_prio, affinity);
   pool->thread_array = (iot_thread_t*) calloc (threads, sizeof (iot_thread_t));
